Reapply "Merge branch 'main' of https://github.com/AnimusindustriesOSX/Pirate-Platformer"
This reverts commit 3e726ad9eced87eb2ec078ba323788624dcfcd9c.

+++ Assets/Scripts/UserInterfaceController.cs
@@ -8,7 +8,7 @@ public class UserInterfaceController : MonoBehaviour
 {
     // Start is called before the first frame update
     [SerializeField] float health = 100;
-    [SerializeField] int insanity;
+    [SerializeField] float insanity;
 
     [SerializeField] int maxHealth = 100;
     [SerializeField] int maxInsanity = 100;
